Move wallet tests into submodules
