+++ tests/lib.rs
@@ -12,7 +12,8 @@ use {
     net::TcpListener,
     path::Path,
     process::{Child, Command, Stdio},
-    str, thread,
+    str::{self, FromStr},
+    thread,
     time::Duration,
   },
   tempfile::TempDir,
